chore: move common types into doppler types and clean AdvancePayOptions
