+++ src/services/doppler-legacy-client.doubles.ts
@@ -85,6 +85,34 @@ export class HardcodedDopplerLegacyClient implements DopplerLegacyClient {
         ],
       },
       {
+        type: 'monthly-deliveries',
+        id: 14,
+        name: '700000-EMAILS-PLUS',
+        emailsByMonth: 700000,
+        extraEmailPrice: 0.00087,
+        fee: 610,
+        featureSet: 'plus',
+        features: ['emailParameter'],
+        billingCycleDetails: [
+          { id: 54, idPlan: 11, paymentType: 'CC', discountPercentage: 0, billingCycle: 'monthly' },
+          {
+            id: 56,
+            idPlan: 14,
+            paymentType: 'CC',
+            discountPercentage: 5,
+            billingCycle: 'quarterly',
+          },
+          {
+            id: 58,
+            idPlan: 14,
+            paymentType: 'CC',
+            discountPercentage: 15,
+            billingCycle: 'half-yearly',
+          },
+          { id: 60, idPlan: 14, paymentType: 'CC', discountPercentage: 25, billingCycle: 'yearly' },
+        ],
+      },
+      {
         type: 'subscribers',
         id: 19,
         name: '2500-SUBSCRIBERS-STANDARD',
